requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.32
botocore==1.34.32
Brotli @ file:///D:/bld/brotli-split_1648883782420/work
certifi==2023.11.17
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
cryptography==42.0.2
datasets==2.17.0
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
diffusers @ file:///home/conda/feedstock_root/build_artifacts/diffusers_1707822915454/work
dill==0.3.8
distro==1.8.0
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.61.0
django-environ==0.11.2
django-rest-authtoken==2.1.4
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1698714947081/work
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1704484084177/work
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1703269254275/work
install==1.3.5
ipykernel @ file:///D:/bld/ipykernel_1707326433944/work
ipython==8.18.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1655912915439/work
jmespath==1.0.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///D:/bld/jupyter_core_1704727156030/work
lightning-utilities==0.10.0
Markdown==3.5.2
MarkupSafe @ file:///D:/bld/markupsafe_1654706079245/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
maturin==1.3.2
mkl-service==2.4.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
multidict==6.0.4
multiprocess==0.70.16
mysqlclient==2.2.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
numpy @ file:///D:/bld/numpy_1653325554875/work
oauthlib==3.2.2
openai==0.28.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.2.0
param==2.0.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
peft==0.7.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
prompt-toolkit==3.0.41
protobuf==4.21.12
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==2.5.2
pydantic_core==2.14.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytorch-lightning==2.1.3
pytz==2023.4
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
regex @ file:///C:/b/abs_d5e2e5uqmr/croot/regex_1696515472506/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
s3transfer==0.10.0
safetensors==0.4.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
sqlparse==0.4.4
stability-sdk==0.8.5
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tf-keras==2.15.0
tokenizers==0.15.0
tomli==2.0.1
torch @ file:///C:/b/abs_6fueooay2f/croot/pytorch-select_1707342446212/work
torchmetrics==1.2.1
torchvision==0.17.0
tornado @ file:///D:/bld/tornado_1656937966227/work
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==3.0.1
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
xxhash==3.4.1
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
